Move compiling/caching routelines, stops, trips, stop_times from traffic_ops to hqta and use this
