+++ high_quality_transit_areas/logs/compile_operators_data.log
@@ -1,0 +1,6 @@
+2022-09-08 14:34:31.167 | INFO     | __main__:<module>:62 - Analysis date: 2022-08-17
+2022-09-08 14:36:34.247 | INFO     | __main__:grab_selected_date:32 - stops compiled and cached
+2022-09-08 14:37:45.086 | INFO     | __main__:grab_selected_date:40 - trips compiled and cached
+2022-09-08 14:40:09.285 | INFO     | __main__:grab_selected_date:48 - routelines compiled and cached
+2022-09-08 14:42:09.749 | INFO     | __main__:grab_selected_date:56 - stop times compiled and cached
+2022-09-08 14:42:09.750 | INFO     | __main__:<module>:69 - execution time: 0:07:38.581752
