+++ traffic_ops/prep_data.py
@@ -12,50 +12,14 @@ import datetime
 import geopandas as gpd
 import pandas as pd
 
-from siuba import *
-from typing import Literal
-
 from shared_utils import geography_utils, gtfs_utils, utils, rt_dates
 
-ANALYSIS_DATE = gtfs_utils.format_date(rt_dates.DATES["jul2022"])
+ANALYSIS_DATE = gtfs_utils.format_date(rt_dates.DATES["aug2022"])
 
 GCS = "gs://calitp-analytics-data/data-analyses/"
 TRAFFIC_OPS_GCS = f"{GCS}traffic_ops/"
 COMPILED_CACHED_GCS = f"{GCS}rt_delay/compiled_cached_views/"
 DATA_PATH = "./data/"
-
-
-def grab_selected_date(selected_date: str):
-    """
-    Create the cached files for stops, trips, stop_times, routes, and route_info
-    """
-    
-    gtfs_utils.all_routelines_or_stops_with_cached(
-        dataset = "stops",
-        analysis_date = selected_date,
-        export_path = COMPILED_CACHED_GCS
-    )
-    
-    gtfs_utils.all_trips_or_stoptimes_with_cached(
-        dataset = "trips",
-        analysis_date = selected_date,
-        export_path = COMPILED_CACHED_GCS
-    )
-    
-    gtfs_utils.all_routelines_or_stops_with_cached(
-        dataset = "routelines",
-        analysis_date = selected_date,
-        export_path = COMPILED_CACHED_GCS
-    )
-    
-    gtfs_utils.all_trips_or_stoptimes_with_cached(
-        dataset = "st",
-        analysis_date = selected_date,
-        export_path = COMPILED_CACHED_GCS
-    )
-    
-    # stops, trips, stop_times, and routes save directly to GCS already
-        
 
 def grab_amtrak(selected_date: datetime.date | str
                ) -> tuple[gpd.GeoDataFrame, pd.DataFrame, gpd.GeoDataFrame]:
@@ -191,7 +155,6 @@ def concatenate_amtrak(
 
 
 def create_local_parquets(selected_date):
-    grab_selected_date(selected_date) 
     grab_amtrak(selected_date)
     concatenate_amtrak(selected_date, COMPILED_CACHED_GCS)
